refactor(email-worker): reorganize into providers/utils, add dotenv, add dev script
- Move SES/Resend senders into providers/ directory
- Move builder and local-runner into utils/ directory
- Rename email-worker.service to send-email
- Add dotenv loading to local-runner for local dev
- Add dev:email-worker npm script
- Add SendEmailPayload type in types.d.ts
