+++ src/workers/email/config.ts
@@ -1,3 +1,4 @@
 export const FROM_EMAIL = 'SVM | ABNMO <naoresponda@abnmo.org>';
+
 export const SUPPORT_WHATSAPP = '(75) 99980-1515';
 export const SUPPORT_EMAIL = 'contato@abnmo.org';
