+++ src/workers/email/types.d.ts
@@ -1,0 +1,5 @@
+export interface SendEmailPayload {
+  to: string;
+  subject: string;
+  html: string;
+}
